+++ sidebar.php
@@ -3,16 +3,15 @@
 /**
  * The main sidebar the theme uses.
  *
- * @package WordPressThemeBoilerplate
- * @author Christopher Lamm <chris@theantichris.com>
- * @since 1.0.0
+ * @package    WordPress
+ * @subpackage WordPressThemeBoilerplate
+ * @author     Christopher Lamm <chris@theantichris.com>
+ * @since      1.0.0
  *
  */
 
 ?>
 
 <div id="sidebar">
-
-	<?php dynamic_sidebar( 'Sidebar Widgets' ); ?>
-
+	<?php dynamic_sidebar( 'sidebar' ); ?>
 </div>